+++ test/test_psf.jl
@@ -5,7 +5,6 @@ using Celeste.BivariateNormals
 using Celeste.PSF
 using Celeste.SensitiveFloats
 
-import Celeste.Util
 import Celeste.SDSSIO
 
 using ForwardDiff
@@ -75,7 +74,7 @@ function test_psf_fit()
     # sig_sf_vec = Array(GalaxySigmaDerivs{NumType}, K);
     #
     # for k = 1:K
-    #   sigma_vec[k] = Util.get_bvn_cov(psf_params[k][psf_ids.e_axis],
+    #   sigma_vec[k] = BivariateNormals.get_bvn_cov(psf_params[k][psf_ids.e_axis],
     #                                   psf_params[k][psf_ids.e_angle],
     #                                   psf_params[k][psf_ids.e_scale])
     #   sig_sf_vec[k] = GalaxySigmaDerivs(
@@ -101,7 +100,7 @@ function test_psf_fit()
 
   sigma_vec = Array(Matrix{Float64}, K);
   for k = 1:K
-    sigma_vec[k] = Util.get_bvn_cov(psf_params[k][psf_ids.e_axis],
+    sigma_vec[k] = BivariateNormals.get_bvn_cov(psf_params[k][psf_ids.e_axis],
                                     psf_params[k][psf_ids.e_angle],
                                     psf_params[k][psf_ids.e_scale])
   end
